+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Added AsciiLettersOnlyTransformer to transform a string to ASCII letters.
 - Added DigitsOnlyTransformer to transform a string to digits.
 - Added LettersOnlyTransformer to transform a string to Unicode letters.
+- Added LowercaseTransformer to transform a string to lowercase.
 
 ### Changed